requirements.txt
langgraph>=0.0.40
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pytest>=7.0.0
typing-extensions>=4.0.0